Breakpoint,  large amount of changes.
Basically I put a new page for creating a DZ using roberts code, added
colors and a info tab. Extended polygon and circle to add a
creator/type/and info on it.
Built filters out for danger zones, and stored them, fixed a weird bug
with the listeners.

TODO — Actually store the danger zones on the device, basically need to
implement the Get and Set function in data manager.